// src/app/interfaces/iaccount.ts
export interface Iaccount {
  account_id?: any;
  username?: string;
  password?: string;
  email?: string;
  firstname?: string;
  secondname?: string;
  lastname?: string;
  picture?: string;
  role_id?: any;
  role?: any;
  information?: string;
  worktitle?: string;
  created_date?: string;
  end_date?: string;
  profile_id?: any;
  profile?: any;
  status?: string;
  instructor_id?: any;
  onlineStatus?: any;
}